+++ app/src/main/java/com/hyeeyoung/wishboard/repository/cart/CartRepository.kt
@@ -3,7 +3,7 @@ package com.hyeeyoung.wishboard.repository.cart
 import com.hyeeyoung.wishboard.model.cart.CartItem
 
 interface CartRepository {
-    suspend fun addToCart(token: String, itemId: Int): Boolean
-    suspend fun removeToCart(token: String, itemId: Int): Boolean
+    suspend fun addToCart(token: String, itemId: Long): Boolean
+    suspend fun removeToCart(token: String, itemId: Long): Boolean
     suspend fun fetchCartList(token: String): List<CartItem>?
 }
